+++ src/demo/Demo.vue
@@ -1,9 +1,21 @@
 <script setup>
+import { defineAsyncComponent } from 'vue'
+import { demos } from './utils/mocks.js'
 
+const TvSettings = defineAsyncComponent(/* webpackChunkName: "tvSettings" */() => import('../components/TvSettings.vue'))
 </script>
 
 <template>
-  <h1>Demo TvSettings</h1>
+  <TvDemo
+    :component="TvSettings"
+    :variants="demos"
+    hide-background
+    component-name="TvSettings"
+    npm-install="@todovue/tv-settings"
+    source-link="https://github.com/TODOvue/tv-settings"
+    url-clone="https://github.com/TODOvue/tv-settings.git"
+    version="1.0.0"
+  />
 </template>
 
 <style scoped>