+++ src/demo/utils/demos/bottom.vue
@@ -1,0 +1,12 @@
+<template>
+  <TvSettings direction="bottom">
+    <template #default>
+      <p>Default bottom panel.</p>
+    </template>
+  </TvSettings>
+</template>
+
+<script setup>
+import { TvSettings } from '@todovue/tv-settings';
+import '@todovue/tv-settings/style.css';
+</script>